server/requirements.txt
fastapi
uvicorn
pydantic
openai
sqlalchemy
python-dotenv
psycopg2-binary
